--- src/main/java/faang/school/urlshortenerservice/ServiceTemplateApplication.java
@@ -1,20 +1,0 @@
-package faang.school.urlshortenerservice;
-
-import org.springframework.boot.Banner;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-@EnableCaching
-@EnableFeignClients("school.faang.servicetemplate.client")
-public class ServiceTemplateApplication {
-    public static void main(String[] args) {
-        new SpringApplicationBuilder(ServiceTemplateApplication.class)
-                .bannerMode(Banner.Mode.OFF)
-                .run(args);
-    }
-}
