--- src/main/java/faang/school/urlshortenerservice/client/FeignConfig.java
@@ -1,15 +1,0 @@
-package faang.school.urlshortenerservice.client;
-
-
-import faang.school.urlshortenerservice.config.context.UserContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfig {
-
-    @Bean
-    public FeignUserInterceptor feignUserInterceptor(UserContext userContext) {
-        return new FeignUserInterceptor(userContext);
-    }
-}
