--- src/main/java/faang/school/urlshortenerservice/encoder/BaseEncoder.java
@@ -1,8 +1,0 @@
-package faang.school.urlshortenerservice.encoder;
-
-import java.util.List;
-
-public interface BaseEncoder {
-
-    List<String> encode(List<Long> numbers);
-}
