--- src/main/java/faang/school/urlshortenerservice/exception/UrlNotFoundException.java
@@ -1,8 +1,0 @@
-package faang.school.urlshortenerservice.exception;
-
-public class UrlNotFoundException extends RuntimeException {
-
-    public UrlNotFoundException(String message) {
-        super(message);
-    }
-}
