--- src/main/java/faang/school/urlshortenerservice/exception/error/ErrorResponse.java
@@ -1,7 +1,0 @@
-package faang.school.urlshortenerservice.exception.error;
-
-import lombok.Builder;
-
-@Builder
-public record ErrorResponse(int code, String message) {
-}
