--- src/main/java/faang/school/urlshortenerservice/model/dto/UrlDto.java
@@ -1,17 +1,0 @@
-package faang.school.urlshortenerservice.model.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UrlDto {
-
-    @NotBlank(message = "URL must not be empty")
-    @Pattern(regexp = "^(http|https)://.*$", message = "URL must start with http:// or https://")
-    private String url;
-}
