+++ src/main/java/ru/corporatiox/urlshortenerservice/ServiceTemplateApplication.java
@@ -1,0 +1,20 @@
+package ru.corporatiox.urlshortenerservice;
+
+import org.springframework.boot.Banner;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+@EnableCaching
+@EnableFeignClients("school.faang.servicetemplate.client")
+public class ServiceTemplateApplication {
+    public static void main(String[] args) {
+        new SpringApplicationBuilder(ServiceTemplateApplication.class)
+                .bannerMode(Banner.Mode.OFF)
+                .run(args);
+    }
+}
