+++ src/main/java/ru/corporatiox/urlshortenerservice/client/FeignConfig.java
@@ -1,0 +1,15 @@
+package ru.corporatiox.urlshortenerservice.client;
+
+
+import ru.corporatiox.urlshortenerservice.config.context.UserContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class FeignConfig {
+
+    @Bean
+    public FeignUserInterceptor feignUserInterceptor(UserContext userContext) {
+        return new FeignUserInterceptor(userContext);
+    }
+}
