+++ src/main/java/ru/corporatiox/urlshortenerservice/client/FeignUserInterceptor.java
@@ -1,0 +1,17 @@
+package ru.corporatiox.urlshortenerservice.client;
+
+import ru.corporatiox.urlshortenerservice.config.context.UserContext;
+import feign.RequestInterceptor;
+import feign.RequestTemplate;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class FeignUserInterceptor implements RequestInterceptor {
+
+    private final UserContext userContext;
+
+    @Override
+    public void apply(RequestTemplate template) {
+        template.header("x-user-id", String.valueOf(userContext.getUserId()));
+    }
+}
