+++ src/main/java/ru/corporatiox/urlshortenerservice/encoder/BaseEncoder.java
@@ -1,0 +1,8 @@
+package ru.corporatiox.urlshortenerservice.encoder;
+
+import java.util.List;
+
+public interface BaseEncoder {
+
+    List<String> encode(List<Long> numbers);
+}
