+++ src/main/java/ru/corporatiox/urlshortenerservice/exception/UrlNotFoundException.java
@@ -1,0 +1,8 @@
+package ru.corporatiox.urlshortenerservice.exception;
+
+public class UrlNotFoundException extends RuntimeException {
+
+    public UrlNotFoundException(String message) {
+        super(message);
+    }
+}
