+++ src/main/java/ru/corporatiox/urlshortenerservice/exception/error/ErrorResponse.java
@@ -1,0 +1,7 @@
+package ru.corporatiox.urlshortenerservice.exception.error;
+
+import lombok.Builder;
+
+@Builder
+public record ErrorResponse(int code, String message) {
+}
