+++ src/main/java/ru/corporatiox/urlshortenerservice/model/entity/Hash.java
@@ -1,0 +1,28 @@
+package ru.corporatiox.urlshortenerservice.model.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@Entity
+@Table(name = "hash")
+public class Hash {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "hash", nullable = false, unique = true, length = 7)
+    private String hash;
+
+    public Hash(String hash) {
+        this.hash = hash;
+    }
+}
