+++ src/main/java/ru/corporatiox/urlshortenerservice/repository/hash/HashRepository.java
@@ -1,0 +1,24 @@
+package ru.corporatiox.urlshortenerservice.repository.hash;
+
+import ru.corporatiox.urlshortenerservice.model.entity.Hash;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface HashRepository extends JpaRepository<Hash, Long> {
+
+    @Query(nativeQuery = true, value = """
+            SELECT nextval('unique_number_seq') FROM generate_series(1, :maxRange);
+            """)
+    List<Long> getUniqueNumbers(long maxRange);
+
+    @Query(nativeQuery = true, value = """
+            DELETE FROM hash WHERE id IN (
+                SELECT id FROM hash ORDER BY id ASC LIMIT :amount
+            ) RETURNING *
+            """)
+    List<Hash> getHashBatch(long amount);
+}
