--- src/test/java/faang/school/urlshortenerservice/ServiceTemplateApplicationTests.java
@@ -1,12 +1,0 @@
-package faang.school.urlshortenerservice;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-
-class ServiceTemplateApplicationTests {
-    @Test
-    void contextLoads() {
-        Assertions.assertThat(40 + 2).isEqualTo(42);
-    }
-}
