--- src/test/java/faang/school/urlshortenerservice/scheduler/CleanerSchedulerTest.java
@@ -1,4 +1,0 @@
-package faang.school.urlshortenerservice.scheduler;
-
-public class CleanerSchedulerTest {
-}
