+++ src/test/java/ru/corporatiox/urlshortenerservice/ServiceTemplateApplicationTests.java
@@ -1,0 +1,12 @@
+package ru.corporatiox.urlshortenerservice;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+
+class ServiceTemplateApplicationTests {
+    @Test
+    void contextLoads() {
+        Assertions.assertThat(40 + 2).isEqualTo(42);
+    }
+}
