+++ src/test/java/ru/corporatiox/urlshortenerservice/scheduler/CleanerSchedulerTest.java
@@ -1,0 +1,4 @@
+package ru.corporatiox.urlshortenerservice.scheduler;
+
+public class CleanerSchedulerTest {
+}
